+++ src/WhatsappApi/Endpoints.php
@@ -24,6 +24,7 @@ class Endpoints
     const GET_BUSINESS_PROFILE = '{phone_number_id}/whatsapp_business_profile';
     const CONFIGURE_WEBHOOK = '{phone_number_id}/webhooks';
     const MANAGE_PHONE_NUMBER = '{phone_number_id}';
+    const OFFICIAL_BUSINESS_ACCOUNT = '{phone_number_id}/official_business_account';
     const BLOCK_USERS = '{phone_number_id}/block_users';
     const UNBLOCK_USERS = '{phone_number_id}/block_users';
     const LIST_BLOCKED_USERS = '{phone_number_id}/block_users';